Put back the old formatting engine, but opt to use the new one in code
